Add files via upload
These macros measure the minimum wall thickness of open or closed objects from the perspective of each pixel on the interior or exterior of a wall. If the wall is broken (incomplete) there is a version that uses an interior object to define the open and closed regions (this also allows a measurement of the percent of wall that is open). A detailed explanation of the two macros is provided in the supplied presentation (PDF and pptx files provided).